Working mockup of press.
Going to try out the current settings for the press that are outlined in this code. While trying to tune the PID by hand I ended with a kP value of 5000 and the other two set to 0 for the time being. I will most likely write a PID auto-tune function so that the values can be updated at a later date.